wip: To all operations with db in controllers and try catch

+++ FitnessFusion-WebProject/FitnessFusion/Controllers/MealController.cs
@@ -3,6 +3,7 @@
     using Microsoft.AspNetCore.Mvc;
     using Services.Data.Interfaces;
     using Web.ViewModels.Meal;
+    using static Common.NotificationMessagesConstant;
 
     public class MealController : Controller
     {
@@ -34,41 +35,106 @@
         {
             if(!ModelState.IsValid)
             {
-                return View();
+                return View(model);
             }
 
-            await mealService.AddMealAsync(model);
+            try
+            {
+                await mealService.AddMealAsync(model);
 
-            return RedirectToAction("All", "Meal");
+                return RedirectToAction("All", "Meal");
+            }
+            catch (Exception)
+            {
+                return GeneralError();
+            }
         }
 
         [HttpPost]
         public async Task<IActionResult> Delete(string id)
         {
-            await mealService.DeleteMealAsync(id);
+            var mealExist = await mealService.IsMealExistByIdAsync(id);
 
-            return RedirectToAction("All", "Meal");
+            if (!mealExist)
+            {
+                TempData[ErrorMessage] = "Meal with provided id does not exist! Please try again!";
+
+                return RedirectToAction("All");
+            }
+
+            try
+            {
+                await mealService.DeleteMealAsync(id);
+
+                return RedirectToAction("All", "Meal");
+            }
+            catch (Exception)
+            {
+                return GeneralError();
+            }
         }
 
         [HttpGet]
         public async Task<IActionResult> Edit(string id)
         {
-            var meal = await mealService.FindMealAsync(id);
+            var mealExist = await mealService.IsMealExistByIdAsync(id);
 
-            return View(meal);
+            if (!mealExist)
+            {
+                TempData[ErrorMessage] = "Meal with provided id does not exist! Please try again!";
+
+                return RedirectToAction("All");
+            }
+
+            try
+            {
+                var meal = await mealService.FindMealAsync(id);
+
+                return View(meal);
+            }
+            catch (Exception)
+            {
+
+                return GeneralError();
+            }
         }
 
         [HttpPost]
         public async Task<IActionResult> Edit(string id, AddMealModel model)
         {
+            var mealExist = await mealService.IsMealExistByIdAsync(id);
+
+            if (!mealExist)
+            {
+                TempData[ErrorMessage] = "Meal with provided id does not exist! Please try again!";
+
+                return RedirectToAction("All");
+            }
+
             if (!ModelState.IsValid)
             {
                 return View(model);
             }
 
-            await mealService.EditMealAsync(id, model);
+            try
+            {
+                await mealService.EditMealAsync(id, model);
 
-            return RedirectToAction("All");
+                return RedirectToAction("All");
+            }
+            catch (Exception)
+            {
+
+                return GeneralError();
+            }
+        }
+
+        private IActionResult GeneralError()
+        {
+            TempData[ErrorMessage] =
+                "Unexpected error occurred! Please try again later or contact administrator";
+
+            return RedirectToAction("Index", "Home");
         }
     }
 }
